+++ mirage/help/usage_flow.py
@@ -17,12 +17,15 @@ Copyright 2017-2018 Shota Shimazu.
 
 from mirage.core import Void
 from mirage.flow import Workflow
-from mirage.help import description
+from mirage.help import description, description_long
 
 class UsageShowWorkFlow(Workflow):
 
     def main(self) -> Void:
-        print(description.usage_doc())
+        if self._option == "--detail":
+            print(description_long.usage_doc())
+        else:
+            print(description.usage_doc())
 
 
 class VersionShowWorkFlow(Workflow):