+++ mirage/system/log.py
@@ -16,9 +16,6 @@ Copyright 2017-2018 Shota Shimazu.
 """
 
 import sys
-import tkinter
-import tkinter.messagebox
-
 
 
 def log(string,
@@ -47,17 +44,13 @@ def log(string,
     elif withConfirm:
         print('\033[31mMirage: ' + str(string) + '\033[0m')
 
-        try:
-            answer = tkinter.messagebox.askyesno("Mirage Confirm", str(string))
-            return answer
-        except:
-            while True:
-                answer = input('\033[32m' + "Please respond with yes or no [Y/N/y/n]" + ' >> \033[0m').lower()
+        while True:
+            answer = input('\033[32m' + "Please respond with yes or no [Y/N/y/n]" + ' >> \033[0m').lower()
 
-                if answer in [ "y", "Y", "yes", "Yes", "YES", "Yeah"]:
-                    return True
-                elif answer in [ "n", "N", "no", "No", "NO", "Nope"]:
-                    return False
+            if answer in [ "y", "Y", "yes", "Yes", "YES", "Yeah"]:
+                return True
+            elif answer in [ "n", "N", "no", "No", "NO", "Nope"]:
+                return False
 
     else:
         print('\033[32mMirage: \033[0m' + str(string))